requirements.txt
boto3==1.19.7
botocore==1.22.7
click==8.0.3
six==1.16.0
python-dotenv==0.21.0
cassandra-driver==3.25.0
